feat: add dropout and flexible output activation
The MLP now applies a dropout layer on the input layer and each hidden
layer.

The activation function of the output layer (if any) can now be passed
directly to the constructor. If none is give, then the results of the
final linear layer is returned.

This also make the `task_type` parameter redundant, which is therefore
removed.