fix: remove reporting of training loss to trial
Optuna threw an user waring before, because the training loss and the
test loss were reported to the optuna study by the trial. But only the
first reported value gets registered, because a study only can optimize
only value. This should be test score. Thus the training loss should not
be reported to the study anyway, aside from preventing the test score to
be reported, if the test score gets reported second.