feat: adjust run method
The Pipelines main function now tries to convert as many step files to
trees and invariants as possible.
No matter wheter the conversion to one representation type fails.

This is because images are now manually
stored as feature data. Filtering for step files with
all representions availablewould mean to delete/filter out certain
images.
Thus it is better trying to convert as much as possible at this point
and equalize the data set later in the ft-mi pipeline.