chore: add gitattribute to normalize line endings
Windows and Linux/Mac use different line endings by default. (CRLF vs
LF)
This leads to diverging of files whenever they are push to the remote
with one type of line ending but locally saved/overwritten with
another line ending.

This gitattribute normalizes line endings when checking files into the
repository but leaves files on the working tree unchanged.

Note: Files checked in before this command need to be renormalized.